added pwa functionality
