+++ client/.gitignore
@@ -1,0 +1,4 @@
+# dependencies
+node_modules
+*/node_modules
+.env
