+++ client/package.json
@@ -2,13 +2,20 @@
   "name": "green-thumb",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
+  "proxy": "http://localhost:3001/",
   "dependencies": {
+    "@apollo/react-hooks": "^4.0.0",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
+    "apollo-boost": "^0.4.9",
+    "graphql": "^15.5.0",
     "react": "^17.0.2",
+    "react-apollo": "^3.1.5",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
+    "serviceworker": "^0.1.7",
     "web-vitals": "^0.2.4",
     "workbox-background-sync": "^5.1.4",
     "workbox-broadcast-update": "^5.1.4",
