+++ server/.gitignore
@@ -1,6 +1,6 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-node_modules
+./node_modules
 */node_modules
 .env
